+++ sdk/deviceregistry/azure-resourcemanager-deviceregistry/src/main/java/com/azure/resourcemanager/deviceregistry/implementation/NamespaceDevicesImpl.java
@@ -11,6 +11,7 @@ import com.azure.core.util.Context;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.deviceregistry.fluent.NamespaceDevicesClient;
 import com.azure.resourcemanager.deviceregistry.fluent.models.NamespaceDeviceInner;
+import com.azure.resourcemanager.deviceregistry.models.DeviceCredentialsRevokeRequest;
 import com.azure.resourcemanager.deviceregistry.models.NamespaceDevice;
 import com.azure.resourcemanager.deviceregistry.models.NamespaceDevices;
 
@@ -63,6 +64,16 @@ public final class NamespaceDevicesImpl implements NamespaceDevices {
         PagedIterable<NamespaceDeviceInner> inner
             = this.serviceClient().listByResourceGroup(resourceGroupName, namespaceName, context);
         return ResourceManagerUtils.mapPage(inner, inner1 -> new NamespaceDeviceImpl(inner1, this.manager()));
+    }
+
+    public void revoke(String resourceGroupName, String namespaceName, String deviceName,
+        DeviceCredentialsRevokeRequest body) {
+        this.serviceClient().revoke(resourceGroupName, namespaceName, deviceName, body);
+    }
+
+    public void revoke(String resourceGroupName, String namespaceName, String deviceName,
+        DeviceCredentialsRevokeRequest body, Context context) {
+        this.serviceClient().revoke(resourceGroupName, namespaceName, deviceName, body, context);
     }
 
     public NamespaceDevice getById(String id) {